Added some new files to 02_programs
1. Added a new script under 02_programs/stata/02_teacher_module_merge.do that cleans some basic variables such as gender and age, runs the fuzzy matching script from python and merges teacher modules to produced the teacher_level.dta used in 04_school_data_merge
