+++ 02_programs/School/Merge_Teacher_Modules/1_clean_raw_data.do
@@ -1,7 +1,7 @@
 /*******************************************************************************
 Purpose: Cleaning all variables in raw data 
 
-Last modified on: 
+Last modified on: 1/30/2024
 By: 
     
 *******************************************************************************/
@@ -18,36 +18,23 @@ matrix drop _all
 global date = c(current_date)
 global username = c(username)
 
-** File paths
+/*Our goal is to clean all variables in all modules before matching teacher names across modules */ 
 
-* set up all globals
-global base "C:\Users\Hersheena\OneDrive\Desktop\Professional\WBG_GEPD_2023\LEGO_Teacher_Paper"
-global data "$base\3_input_data"
-global temp "$base\4_temp_data"
-global final "$base\5_output_data\1_clean_input_data"
-global code "$base\1_code\1_cleaning"
-global log "$base\2_log"
-
-/*Our goal is to clean all variables in all modules before our fuzzy match*/ 
-* The following countries use the same variables
-*Enter the 3 letter abbreviation of the country's data you want to clean here
-*For example, Tchad is TCD
-
-global countries TCD
-																//to identify countries in the global above.e.g 1 if NER and 2 is RWA
-foreach cty_f in $countries {
-* Datasets for PAK_ICT and PAK_KP are saved as PAK
+* Enter the country we are looking at here: PAK_Balochistan
 
 * Now, we start cleaning our datasets
-/* Step 1: Start with roster data */
-use "$data/`cty'/`cty'_teacher_absence.dta", clear
 
-* Dataset should be unique at teacher-id - school_code level
-*isid teachers_id school_code													//6 obs with missing school code in Tchad
+/*********************** Step 1: Start with roster data ************************/
+use "${clone}/01_GEPD_raw_data/School/TEACHERS.dta"
+
+* Dataset should be unique at teacher-id - interview__key level
+rename TEACHERS__id teachers_id
+rename interview__key interview_key
+isid teachers_id interview_key 													
 replace m2saq2=lower(m2saq2)
 
 * Run do file with all value labels
-do "$code/z_value_labels.do"
+do "${clone}/02_programs/School/Merge_Teacher_Modules/z_value_labels.do"
  
 * Sex - Recode sex variable as 1 for female and 0 for male
 recode m2saq3 2=1 1=0
@@ -74,13 +61,14 @@ cap destring urban_rural, replace
 
 * if urban_rural if numeric
 cap recode urban_rural 2=0
-la val urban_rural rural
+cap la val urban_rural rural
 
 * Save file
-save "$final/`cty'/`cty'_teacher_absence.dta", replace
+save "${clone}\01_GEPD_raw_data\School\Cleaned_teacher_modules\teacher_absence.dta", replace
  
-/* Step 2: Clean pedagogy data */
-use "$data/`cty'/`cty'_teacher_pedagogy.dta", clear
+ 
+/************************** Step 2: Clean pedagogy data ****************************/
+/*use "${clone}/GEPD Production Balochistan/01_GEPD_raw_data/School/teacher_pedagogy.dta", clear
 count
 * Data should be unique at m4saq1_number school_code level
 
@@ -103,7 +91,7 @@ cap destring urban_rural, replace
 
 * if urban_rural if numeric
 cap recode urban_rural 2=0
-la val urban_rural rural
+cap la val urban_rural rural
 
 * Label variables 
 cap la var m4scq4_inpt "How many pupils are in the room?" 
@@ -115,51 +103,31 @@ cap la var m4scq11_inpt "How many pupils were not sitting on desks?"
 cap la var m4scq12_inpt "How many students in class as per class list?"
 
 * Save file
-save "$final/`cty'/`cty'_teacher_pedagogy.dta", replace
+save "${clone}\01_GEPD_raw_data\School\Cleaned_teacher_modules\teacher_pedagogy.dta", replace
+*/
+/*************************** Step 3: m3(questionnaire) data **************************/
 
-/* Step 3: m3(questionnaire) data */
-
-use "$data/`cty'/`cty'_teacher_questionnaire.dta", clear
+use "${clone}/01_GEPD_raw_data/School/questionnaire_roster.dta", clear
 * Teacher name is m3sb_troster and teacher id is m3sb_tnumber
-* Data should be unique at m3sb_tnumber-school_code level
+* Data should be unique at m3sb_tnumber- interview key
 
-*Gender
-	tab m7saq10
-	tab m7saq10, nol
-	recode m7saq10 2=1 1=0
-	la val m7saq10 sex
+*isid interview__key m3sb_tnumber
+duplicates tag m3sb_tnumber interview__key, g(tag)
+tab tag
+br if tag ==1
+sort m3sb_tnumber interview__key
+drop tag
+* There are duplicates here
+
+*Rename interview__key
+rename interview__key interview_key
 
 
-*Age - differnt countries have different outliers
-	if `cty'=="NER" {																	//NER
-		* Has an outlier of 1994.
-		sum m3saq6,d	
-		winsor m3saq6, g( m3saq6_w) p(0.001)
-		drop m3saq6
-		rename m3saq6_w m3saq6
-	}
-	if `cty'=="JOR_2023" {																	//Jordan 2023
-		* Has 3 outliers of 247,3140,4015 and 4
-		sum m3saq6,d	
-		winsor m3saq6, g( m3saq6_w) p(0.001)
-		drop m3saq6
-		rename m3saq6_w m3saq6
-		
-		winsor m3saq6, g( m3saq6_w) p(0.003) highonly
-		drop m3saq6
-		rename m3saq6_w m3saq6
-	}
-	if `cty'=="TCD" {																	//Tchad
-		* Has 2 outliers: 1987 and 1991
-		sum m3saq6,d	
-		winsor m3saq6, g( m3saq6_w) h(2) highonly
-		drop m3saq6
-		rename m3saq6_w m3saq6
-	
-	}
-	else {
-		sum m3saq6
-	}
+*Age - there are some outliers here - this removes one obs of 0 and 4 obs that was above 300
+sum m3saq6,d	
+winsor m3saq6, g(m3saq6_w) p(0.006)
+drop m3saq6
+rename m3saq6_w m3saq6
 	
 * Destring urban rural variable and recode
 cap rename rural urban_rural 
@@ -170,21 +138,21 @@ cap destring urban_rural, replace
 
 * if urban_rural if numeric
 cap recode urban_rural 2=0
-la val urban_rural rural
+cap la val urban_rural rural
 
 * Save file
-save "$final/`cty'/`cty'_teacher_questionnaire.dta", replace
+save "${clone}\01_GEPD_raw_data\School\Cleaned_teacher_modules\teacher_questionnaire.dta", replace
 
-/* Step 4: Merge in m5(assessment) data */
-use "$data/`cty'/`cty'_teacher_assessment.dta", clear
+/************************ Step 4: Merge in m5(assessment) data *************************/
+use "${clone}/01_GEPD_raw_data/School/teacher_assessment_answers.dta", clear
 
-* Data should be unique at m5sb_tnumber - school_code
+* Data should be unique at m5sb_tnumber - interview key
 
-*Gender
-	tab m7saq10
-	tab m7saq10, nol
-	recode m7saq10 2=1 1=0
-	la val m7saq10 sex
+* Rename m5sb_tnumber
+rename m5sb_tnum m5sb_tnumber
+
+*Rename interview__key
+rename interview__key interview_key
 
 * Destring urban rural variable and recode
 cap rename rural urban_rural 
@@ -195,10 +163,13 @@ cap destring urban_rural, replace
 
 * if urban_rural if numeric
 cap recode urban_rural 2=0
-la val urban_rural rural
+cap la val urban_rural rural
 
 * Save file
-save "$final/`cty'/`cty'_teacher_assessment.dta", replace
+save "${clone}\01_GEPD_raw_data\School\Cleaned_teacher_modules\teacher_assessment.dta", replace
 
-}
+/************* Run python script for fuzzy matching ****************************/
+python script "${clone}/02_programs/School/Merge_Teacher_Modules/2_teacher_name_matching.py"
+disp "End of teacher name matching"
+
