+++ 02_programs/School/Stata/01_school_run.do
@@ -17,16 +17,17 @@ if _rc != 0 {
 }
 
 
-
-
 *-------------------------------------------------------------------------------
 * Subroutines for this task
 *-------------------------------------------------------------------------------
+* Clean and merge teacher roster module, questionnaire, assessment and classroom observation module
+do "${clone}/02_programs/School/Stata/02_teacher_module_merge.do"
+
 * Import rawdata from .dta files and merge together into four files: school, first grade, fourth grade, teachers
-do "${clone}/02_programs/School/Stata/02_school_data_merge.do"
+do "${clone}/02_programs/School/Stata/03_school_data_merge.do"
 
 * Clean school data
-do "${clone}/02_programs/School/Stata/03_school_data_cleaner.do"
+do "${clone}/02_programs/School/Stata/04_school_data_cleaner.do"
 
 * Anonymize school data
 *do "${clone}/02_programs/School/Stata/04_school_data_anonymizer.do"
